+++ WebStore/Views/Employers/Index.cshtml
@@ -1,14 +1,6 @@
-﻿@model IEnumerable<WebStore.Models.Employer>
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <title>Работники</title>
-</head>
-<body>
+﻿<div class="container">
     <h3>Работники предприятия</h3>
-    <table>
+    <table class="table">
         <thead>
             <tr>
                 <th>ID</th>
@@ -30,5 +22,4 @@
             }
             </thead>
         </table>
-</body>
-</html>
+</div>
