+++ WebStore/Views/Shared/Partial/_Header.cshtml
@@ -30,7 +30,7 @@
 				<div class="row">
 					<div class="col-sm-4">
 						<div class="logo pull-left">
-							<a href="index.html"><img src="images/home/logo.png" alt="" /></a>
+							<a asp-controller="Home" asp-action="Index"><img src="images/home/logo.png" alt="" /></a>
 						</div>
 						<div class="btn-group pull-right">
 							<div class="btn-group">
@@ -83,28 +83,7 @@
 								<span class="icon-bar"></span>
 							</button>
 						</div>
-						<div class="mainmenu pull-left">
-							<ul class="nav navbar-nav collapse navbar-collapse">
-								<li><a href="index.html" class="active">Home</a></li>
-								<li class="dropdown"><a href="#">Магазин<i class="fa fa-angle-down"></i></a>
-                                    <ul role="menu" class="sub-menu">
-                                        <li><a href="shop.html">Каталог</a></li>
-										<li><a href="product-details.html">О товаре</a></li> 
-										<li><a href="checkout.html">Оформить заказ</a></li> 
-										<li><a href="cart.html">Корзина</a></li> 
-										<li><a href="login.html">Вход</a></li> 
-                                    </ul>
-                                </li> 
-								<li class="dropdown"><a href="#">Блог<i class="fa fa-angle-down"></i></a>
-                                    <ul role="menu" class="sub-menu">
-                                        <li><a href="blog.html">Список блогов</a></li>
-										<li><a href="blog-single.html">Блог магазина</a></li>
-                                    </ul>
-                                </li> 
-								<li><a href="404.html">404</a></li>
-								<li><a href="contact-us.html">Контакты</a></li>
-							</ul>
-						</div>
+						<partial name="Partial/_MainMenu"/>
 					</div>
 					<div class="col-sm-3">
 						<div class="search_box pull-right">
